+++ langs/cpp/maker.sh
@@ -2,6 +2,11 @@ NAME=$2
 DIRLOCALE=$(dirname -- $(readlink -fn -- "$0"))
 
 if [ $1 == "file" ]; then
+	if [ -f $NAME.cpp ]; then
+		echo "The file cNAME.hpp exist, please remove to proced"
+		exit 1
+	fi
+
 	cp $DIRLOCALE/main.cpp ./$NAME.cpp
 fi
 
@@ -11,6 +16,11 @@ if [ $1 == "dir" ]; then
 fi
 
 if [ $1 == "lib" ]; then
+	if [ -f $NAME.hpp ]; then
+		echo "The file $NAME.hpp exist, please remove to proced"
+		exit 1
+	fi
+
 	echo -e "#ifndef "$NAME"_HPP_INCLUDED"    >> $NAME.hpp
 	echo -e "#define "$NAME"_HPP_INCLUDED"    >> $NAME.hpp
 	echo -e "#include <iostream>"             >> $NAME.hpp
