+++ langs/lua/maker.sh
@@ -2,6 +2,11 @@ NAME=$2
 DIRLOCALE=$(dirname -- $(readlink -fn -- "$0"))
 
 if [ $1 == "file" ]; then
+	if [ -f $NAME.lua ]; then
+		echo "The file $NAME.lua exist, please remove to proced"
+		exit 1
+	fi
+
 	cp $DIRLOCALE/main.lua ./$NAME.lua
 fi
 
@@ -10,5 +15,10 @@ if [ $1 == "dir" ]; then
 fi
 
 if [ $1 == "lib" ]; then
+	if [ -f $NAME.lua ]; then
+		echo "The file $NAME.lua exist, please remove to proced"
+		exit 1
+	fi
+
 	cp $DIRLOCALE/lib.lua ./$NAME.lua
 fi
