+++ langs/nodejs/maker.sh
@@ -2,6 +2,11 @@ NAME=$2
 DIRLOCALE=$(dirname -- $(readlink -fn -- "$0"))
 
 if [ $1 == "file" ]; then
+	if [ -f $NAME.js ]; then
+		echo "The file $NAME.js exist, please remove to proced"
+		exit 1
+	fi
+
 	cp $DIRLOCALE/main.js ./$NAME.js
 fi
