+++ langs/python/maker.sh
@@ -2,6 +2,11 @@ NAME=$2
 DIRLOCALE=$(dirname -- $(readlink -fn -- "$0"))
 
 if [ $1 == "file" ]; then
+	if [ -f $NAME.py ]; then
+		echo "The file $NAME.py exist, please remove to proced"
+		exit 1
+	fi
+
 	cp $DIRLOCALE/main.py ./$NAME.py
 fi
 
@@ -10,6 +15,11 @@ if [ $1 == "dir" ]; then
 fi
 
 if [ $1 == "lib" ]; then
+	if [ -d $NAME ]; then
+		echo "The directory $NAME exist, please remove to proced"
+		exit 1
+	fi
+
 	mkdir -p $NAME
 	cp $DIRLOCALE/__init__.py $NAME/
 fi
