+++ langs/web/maker.sh
@@ -2,6 +2,11 @@ NAME=$2
 DIRLOCALE=$(dirname -- $(readlink -fn -- "$0"))
 
 if [ $1 == "file" ]; then
+	if [ -f $NAME.html ]; then
+		echo "The file $NAME.html exist, please remove to proced"
+		exit 1
+	fi
+
 	cp $DIRLOCALE/index.html ./$NAME.html
 fi
